# crates/sdk/core/specs/gateway-openapi.yaml
openapi: 3.1.0
info:
  title: Ryu Gateway API
  version: "1.0.0"
  description: |
    The **Ryu Gateway** is the OpenAI-compatible control layer that governs every
    model call: routing, fallback, rate-limiting, firewall/PII-DLP, budgets, evals,
    and audit. Point any OpenAI-compatible client at the gateway by swapping its
    base URL; the request surface below is a superset of the OpenAI API plus Ryu's
    governance, eval, and audit endpoints.

    Authentication is via a bearer token (`Authorization: Bearer <api-key>`). When
    the gateway runs with `auth.require_auth = false`, anonymous requests are
    allowed. Admin endpoints (`/v1/config`, `/v1/audit`) additionally require the
    master key or a loopback peer.
servers:
  - url: http://localhost:7981
    description: Local gateway (default bind)
security:
  - bearerAuth: []
...
paths:
  /v1/chat/completions:
    post:
      tags: [Inference]
      summary: Create a chat completion
      description: |
        OpenAI-compatible chat completion. Set `stream: true` to receive a
        Server-Sent Events stream of `chat.completion.chunk` objects. Per-request
        governance is driven by optional `x-ryu-*` headers (user/agent identity,
        per-agent provider/model slot overrides, session correlation).
      operationId: createChatCompletion
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/AgentIdHeader"
        - name: x-ryu-slot-chat-provider
          in: header
          required: false
          description: Per-agent chat provider override.
          schema: { type: string }
        - name: x-ryu-slot-chat-model
          in: header
          required: false
          description: Per-agent chat model override.
          schema: { type: string }
        - name: x-ryu-session-id
          in: header
          required: false
          description: Core conversation id, recorded in the audit log for correlation.
          schema: { type: string }
        - name: x-ryu-prompt-cache
          in: header
          required: false
          description: |
            Provider prompt-cache mode for this request, overriding the node's
            `[prompt_cache].mode`. `off` injects nothing; `auto` lets the provider
            place the breakpoint; `explicit` places breakpoints on the system
            prompt and trailing turns. Cache markers already present in the request
            body always win over any of these. Ignored when the node sets
            `allow_request_override = false`. Unset ⇒ the node's config decides.
          schema: { type: string, enum: [off, auto, explicit] }
        - name: x-ryu-prompt-cache-ttl
          in: header
          required: false
          description: |
            Requested `cache_control.ttl`. `1h` bills more to write than the 5-minute
            default, so it is opt-in. Unset ⇒ the provider default.
          schema: { type: string, enum: ["5m", "1h"] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: A chat completion (or an SSE stream when `stream` is true).
          headers:
            x-request-id:
              description: Gateway request id.
              schema: { type: string }
            x-provider:
              description: Upstream provider that served the request.
              schema: { type: string }
            x-cache:
              description: Exact-cache result.
              schema: { type: string, enum: [HIT, MISS] }
            x-degraded:
              description: Set to `fallback:<provider>` when served by a fallback.
              schema: { type: string }
            x-ryu-prompt-cache:
              description: |
                What the prompt-cache stage did. `injected` markers were added;
                `caller` the body carried a root-level marker so the caller is
                driving; `caller-breakpoints` a marker already sat inside
                `messages`/`tools` and suppressed injection (a stray one from a
                turn hook or a replayed transcript looks the same as a deliberate
                one, hence the separate value); `automatic` the model family caches
                with no marker; `below-minimum` the prompt was too small to be
                cacheable; `disabled` injection is off (also the value when this
                gateway's own response cache answered, i.e. `x-cache: HIT`).
              schema:
                type: string
                enum:
                  [
                    injected,
                    caller,
                    caller-breakpoints,
                    automatic,
                    below-minimum,
                    disabled,
                  ]
            x-ryu-cache-read:
              description: |
                Prompt tokens the provider served from its cache. Non-streaming only —
                on a stream the counts arrive after the headers are sent.
              schema: { type: integer }
            x-ryu-cache-write:
              description: |
                Prompt tokens the provider wrote into its cache. Billed above the
                normal input rate, so it is reported apart from the read count.
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                type: string
                description: "SSE stream of chat.completion.chunk objects (one per `data:` line)."
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/images/generations:
    post:
      tags: [Inference]
      summary: Create an image
      operationId: createImage
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/AgentIdHeader"
        - name: x-ryu-slot-image-provider
          in: header
          required: false
          schema: { type: string }
        - name: x-ryu-slot-image-model
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageGenerationRequest"
      responses:
        "200":
          description: Generated image(s).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerationResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key issued by the gateway, or the master key for admin endpoints.
  parameters:
    UserIdHeader:
      name: x-ryu-user-id
      in: header
      required: false
      description: Per-user budget identity.
      schema: { type: string }
    AgentIdHeader:
      name: x-ryu-agent-id
      in: header
      required: false
      description: Per-agent budget identity.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Rate limit or budget stop reached.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: string }
    ChatMessage:
      type: object
      properties:
        role: { type: string, enum: [system, user, assistant, tool] }
        content: { type: string }
      required: [role, content]
    ChatCompletionRequest:
      type: object
      properties:
        model: { type: string, description: Routed via the gateway model map. }
        messages:
          type: array
          items: { $ref: "#/components/schemas/ChatMessage" }
        stream: { type: boolean, default: false }
        temperature: { type: number, format: float }
        top_p: { type: number, format: float }
        max_tokens: { type: integer }
      required: [model, messages]
      additionalProperties: true
    ChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, const: chat.completion }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message: { $ref: "#/components/schemas/ChatMessage" }
              finish_reason: { type: string }
        usage: { $ref: "#/components/schemas/Usage" }
    Usage:
      type: object
      properties:
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
    ImageGenerationRequest:
      type: object
      properties:
        model: { type: string }
        prompt: { type: string }
        "n": { type: integer, default: 1 }
        size: { type: string, examples: ["1024x1024"] }
        quality: { type: string, enum: [standard, hd] }
        response_format: { type: string, enum: [url, b64_json] }
      required: [prompt]
      additionalProperties: true
    ImageGenerationResponse:
      type: object
      properties:
        created: { type: integer }
        data:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              b64_json: { type: string }
...